Remove base class for ClickHouse connector test
